+++ src/event_view/event_item.h
@@ -93,6 +93,7 @@ public:
   [[nodiscard]] QString getTitle() const;
   [[nodiscard]] QString getClientName() const;
   [[nodiscard]] std::optional<double> cost() const;
+  [[nodiscard]] int64_t paymentStatusId() const;
 
   /**
    * @brief Returns the unique identifier of the event.
@@ -109,6 +110,7 @@ public:
   void setTitle(const QString &title);
   void setClientName(const QString &clientName);
   void setCost(std::optional<double> cost);
+  void setPaymentStatusId(int64_t paymentStatusId);
 
   /**
    * @brief Sets the start time of the event.
@@ -166,6 +168,7 @@ private:
   QString mTitle;             ///< The title of the event.
   QString mClientName;        ///< Linked client name for work items.
   std::optional<double> mCost;///< Optional work event cost.
+  int64_t mPaymentStatusId = 1; ///< Payment status for work items.
   QDateTime mStartTime;       ///< The start time of the event.
   QDateTime mEndTime;         ///< The end time of the event.
   unsigned long mId;          ///< Unique identifier of the event.